Refactor AddOrEditTrigger
